+++ commons/src/test/java/com/karuslabs/commons/command/ExceptionsTest.java
@@ -23,6 +23,7 @@
  */
 package com.karuslabs.commons.command;
 
+import com.karuslabs.commons.command.dispatcher.Exceptions;
 import com.mojang.brigadier.LiteralMessage;
 import com.mojang.brigadier.exceptions.CommandSyntaxException;